# docs/swagger.yaml
definitions:
...
  vo.BonusHistory:
    properties:
      address:
        type: string
      associate:
        type: integer
      bonus:
        type: integer
      create_time:
        type: integer
      state:
        type: integer
      type:
        type: integer
    type: object
...
  vo.BonusWithdraw:
    properties:
      address:
        type: string
      apply_time:
        type: integer
      bonus:
        type: integer
      gas:
        type: integer
      id:
        type: integer
      nick_name:
        type: string
      state:
        type: integer
      txhash:
        type: string
    type: object
...
  vo.Pagination:
    properties:
      curPage:
        type: integer
      items: {}
      totalNum:
        type: integer
    type: object
...
  vo.ProcessWithdrawBonusReq:
    properties:
      id:
        type: integer
      txhash:
        type: string
    type: object
...
  vo.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  vo.UserBonus:
    properties:
      address:
        type: string
      balance:
        type: integer
      freeze:
        type: integer
      today_win:
        type: integer
      total_win:
        type: integer
    type: object
...
  vo.WithdrawBonusApplyReq:
    properties:
      address:
        type: string
      bonus:
        type: integer
      nonce:
        type: string
      signed_msg:
        type: string
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /bonus/histories:
    get:
      consumes:
      - application/json
      description: get user bonus history
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      - description: 1：win 2：withdraw
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.BonusHistory'
                        type: array
                    type: object
              type: object
      summary: GetUserBonusHistory
      tags:
      - bonus
  /bonus/my:
    get:
      consumes:
      - application/json
      description: get my bonus
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserBonus'
              type: object
      summary: GetMyBonus
      tags:
      - bonus
  /bonus/my/history:
    get:
      consumes:
      - application/json
      description: get my bonus history
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.BonusHistory'
                        type: array
                    type: object
              type: object
      summary: GetMyBonusHistory
      tags:
      - bonus
  /bonus/withdraw/apply:
    post:
      consumes:
      - application/json
      description: apply to withdraw bonus
      parameters:
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.WithdrawBonusApplyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: WithdrawBonusApply
      tags:
      - bonus
...
  /broadcast:
    get:
      consumes:
      - application/json
      description: get broadcasts
      parameters:
      - description: language
        in: header
        name: lang
        required: true
        type: string
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          type: string
                        type: array
                    type: object
              type: object
      summary: GetBroadcasts
      tags:
      - broadcast
  /broadcast/latest:
    get:
      consumes:
      - application/json
      description: get the latest broadcast
      parameters:
      - description: language
        in: header
        name: lang
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GetLatestBroadcast
      tags:
      - broadcast
...
  /withdraw/histories:
    get:
      consumes:
      - application/json
      description: get  bonus withdraw ,can filter by address and type
      parameters:
      - description: filter address,if empty will return all user
        in: query
        name: address
        type: string
      - description: type 0：unProcess 1:processing 2:processed
        in: query
        name: type
        type: integer
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.BonusWithdraw'
                        type: array
                    type: object
              type: object
      summary: GetWithdraws
      tags:
      - withdraw
  /withdraw/my/history:
    get:
      consumes:
      - application/json
      description: get my bonus withdraw ,can filter by type
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: type 0：unProcess 1:processing 2:processed
        in: query
        name: type
        type: integer
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.BonusWithdraw'
                        type: array
                    type: object
              type: object
      summary: GetMyWithdraws
      tags:
      - withdraw
  /withdraw/process:
    post:
      consumes:
      - application/json
      description: process withdraw bonus
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.ProcessWithdrawBonusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: WithdrawBonusProcess
      tags:
      - withdraw
  /withdraw/report_tx:
    post:
      consumes:
      - application/json
      description: report processed withdraw tx hash
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.ProcessWithdrawBonusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: WithdrawReportHash
      tags:
      - withdraw
swagger: "2.0"